refactor: move resume game functions to own module resume_game.js
