+++ nonograms/scripts/game_logic.js
@@ -33,21 +33,7 @@ function manageHeader(event) {
   }
 
   if (buttonId === 'resume') {
-    slConfig.getLastConfig();
-    audio.setBlockAudio(true);
-    setSaveResumeButton('save');
-    const lastSave = vars.resumeLastGame().game.answerCells;
-    const colTips = lastSave.col_tips;
-    const rowTips = lastSave.row_tips;
-    const colCross = lastSave.col_cross;
-    const rowCross = lastSave.row_cross;
-    setGameByLevel(event, vars.currentLevelId, table[0]);
-    const picArray = new Puzzle(colTips, rowTips).getPicByTips();
-    const crossArray = new Puzzle(colCross, rowCross).getPicByTips();
-    restoreData(['fill'], picArray);
-    restoreData(['cross'], crossArray);
-    fillTable(['fill'], picArray.flat());
-    fillTable(['cross'], crossArray.flat());
+    resumeGame(event, table);
   }
 
   if (buttonId === 'repeat') {
@@ -66,16 +52,6 @@ function manageHeader(event) {
     toggleSidebar(event);
     isOverlayShown ? removeOverlay(event) : setOverlay(event);
   }
-}
-
-function restoreData(sign, dataArray) {
-  dataArray.forEach((row, row_index) => {
-    row.forEach((cell, col_index) => {
-      if (cell === 1) {
-        setAnswer(sign, col_index, row_index);
-      }
-    });
-  });
 }
 
 function setSaveResumeButton(name) {
